Move ExtendedEntity to proper package
